backend/requirements.txt
flask
websockets
pika
PyYAML